# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GenaiError {
  error: string;
}

export type GateStatusStatus =
  (typeof GateStatusStatus)[keyof typeof GateStatusStatus];

export const GateStatusStatus = {
  clear: "clear",
  moderate: "moderate",
  congested: "congested",
  critical: "critical",
} as const;

export interface GateStatus {
  id: string;
  name: string;
  crowdPct: number;
  status: GateStatusStatus;
}

export type TransportLineMode =
  (typeof TransportLineMode)[keyof typeof TransportLineMode];

export const TransportLineMode = {
  rail: "rail",
  shuttle: "shuttle",
  rideshare: "rideshare",
  parking: "parking",
} as const;

export type TransportLineStatus =
  (typeof TransportLineStatus)[keyof typeof TransportLineStatus];

export const TransportLineStatus = {
  normal: "normal",
  delayed: "delayed",
  disrupted: "disrupted",
} as const;

export interface TransportLine {
  name: string;
  mode: TransportLineMode;
  status: TransportLineStatus;
  etaMinutes: number;
}

export type OperationalStateActiveIncident =
  (typeof OperationalStateActiveIncident)[keyof typeof OperationalStateActiveIncident];

export const OperationalStateActiveIncident = {
  none: "none",
  storm: "storm",
  transit_disruption: "transit_disruption",
  crowd_surge: "crowd_surge",
  grid_failure: "grid_failure",
} as const;

export interface OperationalState {
  timestamp: string;
  weatherCondition: string;
  energyLoadPct: number;
  crowdCount: number;
  activeIncident: OperationalStateActiveIncident;
  gates: GateStatus[];
  transport: TransportLine[];
}

export type DirectiveCategory =
  (typeof DirectiveCategory)[keyof typeof DirectiveCategory];

export const DirectiveCategory = {
  crowd: "crowd",
  accessibility: "accessibility",
  transport: "transport",
  sustainability: "sustainability",
  security: "security",
  ticketing: "ticketing",
} as const;

export type DirectiveSeverity =
  (typeof DirectiveSeverity)[keyof typeof DirectiveSeverity];

export const DirectiveSeverity = {
  info: "info",
  watch: "watch",
  critical: "critical",
} as const;

export type DirectiveStatus =
  (typeof DirectiveStatus)[keyof typeof DirectiveStatus];

export const DirectiveStatus = {
  executing: "executing",
  monitoring: "monitoring",
  resolved: "resolved",
} as const;

export interface Directive {
  id: string;
  title: string;
  detail: string;
  category: DirectiveCategory;
  severity: DirectiveSeverity;
  /** @nullable */
  gate: string | null;
  status: DirectiveStatus;
}

export interface OperationsBrief {
  summary: string;
  generatedAt: string;
  directives: Directive[];
}

export interface ConciergeMessageInput {
  /**
   * @minLength 1
   * @maxLength 500
   */
  message: string;
}

export type ConciergeMessageOutputCategory =
  (typeof ConciergeMessageOutputCategory)[keyof typeof ConciergeMessageOutputCategory];

export const ConciergeMessageOutputCategory = {
  navigation: "navigation",
  accessibility: "accessibility",
  transportation: "transportation",
  ticketing: "ticketing",
  general: "general",
} as const;

export interface ConciergeMessageOutput {
  detectedLanguage: string;
  translatedMessage: string;
  reply: string;
  replyTranslation: string;
  category: ConciergeMessageOutputCategory;
}

export type SimulateIncidentBodyType =
  (typeof SimulateIncidentBodyType)[keyof typeof SimulateIncidentBodyType];

export const SimulateIncidentBodyType = {
  none: "none",
  storm: "storm",
  transit_disruption: "transit_disruption",
  crowd_surge: "crowd_surge",
  grid_failure: "grid_failure",
} as const;

export interface SimulateIncidentBody {
  type: SimulateIncidentBodyType;
}

export type UpdateCustomTelemetryBodyActiveIncident =
  (typeof UpdateCustomTelemetryBodyActiveIncident)[keyof typeof UpdateCustomTelemetryBodyActiveIncident];

export const UpdateCustomTelemetryBodyActiveIncident = {
  none: "none",
  storm: "storm",
  transit_disruption: "transit_disruption",
  crowd_surge: "crowd_surge",
  grid_failure: "grid_failure",
  manual: "manual",
} as const;

export type UpdateCustomTelemetryBodyGatesItem = {
  id: string;
  crowdPct: number;
};

export type UpdateCustomTelemetryBodyTransportItemStatus =
  (typeof UpdateCustomTelemetryBodyTransportItemStatus)[keyof typeof UpdateCustomTelemetryBodyTransportItemStatus];

export const UpdateCustomTelemetryBodyTransportItemStatus = {
  normal: "normal",
  delayed: "delayed",
  disrupted: "disrupted",
} as const;

export type UpdateCustomTelemetryBodyTransportItem = {
  name: string;
  etaMinutes: number;
  status: UpdateCustomTelemetryBodyTransportItemStatus;
};

export interface UpdateCustomTelemetryBody {
  weatherCondition?: string;
  energyLoadPct?: number;
  activeIncident?: UpdateCustomTelemetryBodyActiveIncident;
  gates?: UpdateCustomTelemetryBodyGatesItem[];
  transport?: UpdateCustomTelemetryBodyTransportItem[];
}

export type SimulateIncident200 = {
  status: string;
};

export type ResetIncident200 = {
  status: string;
};

export type UpdateCustomTelemetry200 = {
  status: string;
};