# lib/api-zod/src/generated/types/conciergeMessageOutput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConciergeMessageOutputCategory } from "./conciergeMessageOutputCategory";

export interface ConciergeMessageOutput {
  detectedLanguage: string;
  translatedMessage: string;
  reply: string;
  replyTranslation: string;
  category: ConciergeMessageOutputCategory;
}