# lib/api-zod/src/generated/types/directive.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DirectiveCategory } from "./directiveCategory";
import type { DirectiveSeverity } from "./directiveSeverity";
import type { DirectiveStatus } from "./directiveStatus";

export interface Directive {
  id: string;
  title: string;
  detail: string;
  category: DirectiveCategory;
  severity: DirectiveSeverity;
  /** @nullable */
  gate: string | null;
  status: DirectiveStatus;
}