# lib/api-zod/src/generated/types/gateStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GateStatusStatus } from "./gateStatusStatus";

export interface GateStatus {
  id: string;
  name: string;
  crowdPct: number;
  status: GateStatusStatus;
}