# lib/api-zod/src/generated/types/operationsBrief.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Directive } from "./directive";

export interface OperationsBrief {
  summary: string;
  generatedAt: Date;
  directives: Directive[];
}