# lib/api-zod/src/generated/types/transportLine.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransportLineMode } from "./transportLineMode";
import type { TransportLineStatus } from "./transportLineStatus";

export interface TransportLine {
  name: string;
  mode: TransportLineMode;
  status: TransportLineStatus;
  etaMinutes: number;
}