# lib/api-zod/src/generated/types/updateCustomTelemetryBody.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCustomTelemetryBodyActiveIncident } from "./updateCustomTelemetryBodyActiveIncident";
import type { UpdateCustomTelemetryBodyGatesItem } from "./updateCustomTelemetryBodyGatesItem";
import type { UpdateCustomTelemetryBodyTransportItem } from "./updateCustomTelemetryBodyTransportItem";

export interface UpdateCustomTelemetryBody {
  weatherCondition?: string;
  energyLoadPct?: number;
  activeIncident?: UpdateCustomTelemetryBodyActiveIncident;
  gates?: UpdateCustomTelemetryBodyGatesItem[];
  transport?: UpdateCustomTelemetryBodyTransportItem[];
}